# docs/swagger.yaml
definitions:
  client.ProfilesFiles3DResponse:
    properties:
      downloaded:
        type: boolean
      errors:
        items:
          type: string
        type: array
      jpg_path:
        type: string
      jpg_url:
        type: string
      mtl_path:
        type: string
      mtl_url:
        type: string
      obj_path:
        type: string
      obj_url:
        type: string
      ply_path:
        type: string
      ply_url:
        type: string
      project_id:
        type: integer
    type: object
...
  handler.CreateOrUpdateProjectRequest:
    properties:
      address:
        example: 123 Green Energy Street, San Diego, CA
        type: string
      description:
        example: Residential rooftop solar project
        type: string
      name:
        example: Solar Panel Installation
        type: string
      status:
        example: active
        type: string
      user_id:
        example: 1
        type: integer
    required:
    - name
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  handler.LeadResponse:
    properties:
      house_id:
        example: 123
        type: integer
      lead_id:
        example: 42
        type: integer
      success:
        example: true
        type: boolean
    type: object
...
  models.Project:
    properties:
      address:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  service.CreateLead:
    properties:
      consumption:
        items:
          type: integer
        type: array
      hardware_type:
        type: string
      kwh_usage:
        example: 12000
        type: number
      latitude:
        example: 37.7749
        type: number
      longitude:
        example: -122.4194
        type: number
      lse_id:
        type: integer
      mode:
        type: string
      panel_count:
        example: 30
        type: integer
      period:
        type: string
      project_id:
        example: 1
        type: integer
      system_size:
        example: 10.5
        type: number
      target_solar_offset:
        type: integer
      unit:
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@sunready.com
    name: API Support
  description: API for Sun Ready project
  license:
    name: Sun Ready Private License
    url: INTERNAL
  termsOfService: http://swagger.io/terms/
  title: Sun Ready API
  version: "1.0"
paths:
...
  /api/leads:
    get:
      description: Retrieves a paginated list of leads
      parameters:
      - description: Filter by customer ID
        in: query
        name: customer_id
        type: integer
      - description: Filter by creator ID
        in: query
        name: creator_id
        type: integer
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List leads
      tags:
      - leads
    post:
      consumes:
      - application/json
      description: Creates a new lead and optionally initiates 3D model generation
      parameters:
      - description: Lead details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateLead'
      produces:
      - application/json
      responses:
        "201":
          description: Returns lead and house IDs
          schema:
            $ref: '#/definitions/handler.LeadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new lead
      tags:
      - leads
  /api/leads/{id}:
    delete:
      description: Deletes a lead by ID
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a lead
      tags:
      - leads
    get:
      description: Retrieves a lead by its ID
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LeadResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a lead by ID
      tags:
      - leads
    put:
      consumes:
      - application/json
      description: Updates an existing lead
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lead updates
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LeadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update a lead
      tags:
      - leads
...
  /api/projects:
    post:
      consumes:
      - application/json
      description: Creates a new project record in the system.
      parameters:
      - description: Project data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOrUpdateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Project created successfully
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Failed to create project
          schema:
            type: string
      summary: Create a new project
      tags:
      - Projects
  /api/projects/{id}:
    delete:
      description: Deletes a project by its ID.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Project deleted successfully
          schema:
            type: string
        "400":
          description: Invalid project ID
          schema:
            type: string
        "500":
          description: Failed to delete project
          schema:
            type: string
      summary: Delete a project
      tags:
      - Projects
    get:
      description: Retrieves a project by its unique ID.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project found
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Invalid project ID
          schema:
            type: string
        "404":
          description: Project not found
          schema:
            type: string
      summary: Get project by ID
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Updates an existing project by its ID.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated project data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOrUpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated successfully
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Invalid request body or project ID
          schema:
            type: string
        "500":
          description: Failed to update project
          schema:
            type: string
      summary: Update a project
      tags:
      - Projects
  /api/projects/user:
    get:
      description: Retrieves all projects associated with a specific user, with pagination.
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      - description: 'Results limit (default: 20)'
        in: query
        name: limit
        type: integer
      - description: Results offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of projects
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
        "400":
          description: Invalid user ID
          schema:
            type: string
        "500":
          description: Failed to fetch projects
          schema:
            type: string
      summary: List projects by user ID
      tags:
      - Projects
...
  /leads/{id}/mesh-files:
    get:
      consumes:
      - application/json
      description: Retrieves the 3D mesh files associated with a specific lead ID
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.ProfilesFiles3DResponse'
        "400":
          description: Invalid lead ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Lead not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get 3D mesh files for a lead
      tags:
      - Leads
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"